tools/claude-dev/requirements.txt
click>=8.0.0
pyyaml>=6.0
jinja2>=3.0
rich>=13.0
pytest>=7.0
coverage>=7.0

# Optional: for agent integration
# anthropic>=0.18.0
